requirements.txt
flask==3.0.2
pandas==2.2.1
numpy==1.26.4
scikit-surprise==1.1.4
scikit-learn==1.4.1.post1
